Remove base and callbacks from proof_of_concept
The proof of concept is done to have some reasonable
results for the first time... The callbacks made some issues,
so TensorBoard is a matter of future work (but definitely wanted)
and the bases are omitted because this is a binary classifier.
Tests with 3 classes are also matter of future work.