Add Cookiecutter data structure
Cookiecutter data structure is a standard for data science projects,
which is implement through this commit.
All testing capabilities are removed though, because the intend of this
project is not to go into production, but simply to produce models
for later usage in other applications.

However they might be added later.

A lot of code from the proof of concept has been altered,
transfering it into appropriate python modules which are then imported.
This allows for greater flexibility and might safe time when creating
models to come.

The code for data_generation has been imported from the former
data_generation repository (which is deleted now), because the
Cookiecutter template allows for adequate integration.
It is cleaner this way...

Cookiecutter: https://github.com/drivendata/cookiecutter-data-science