Add frame for srp
srp stands for Student Research Project which will represent  the
first stage for deepmech. Target is to develop a model which takes
mechanical symbols as input and gives the correct representation as
a suitable API for other libraries like mec2 to work with.

This report is a paper which introduces the reader into concepts of
machine learning and shows how they are applied to create said model.