Add functioning model for constraints
The tests with YOLO where unsuccessful.
They are included in this commit for further tests.
Problem is that YOLO does not detect nodes in environments
with cluttered background, like constraints.

This should not be. But it is.
Other tests were made with detecting constraints:
- Using their angle
- Trying to get their bounding boxes.
- Not sure what yolo4 was about...
    i think the test was unfinished

I tried to use the first fcn to detect constraints,
by creating a model which detects constraints between
nodes by cropping the image and detecting an arrow from
top-right to bottom-left, by the necessary operations it
can determine which direction was used.

The model is then converted to the tensorflow.js format like
it was previously done with the symbol_detector.

tensorflow.js has horrible implementation for preparing
an image for that case...so the preparation of all the boxes
takes the most time. This pipeline has to be optimized.

But for now it is a working model which detects
nodes AND constraints.