Update node detector
Node detector was updated to not detect mec2 symbols and is now
specialized on white images on black background.
Furthermore attempts have been made to include a precompiled FCN into
TensorflowJS but those were unsuccessful.
Beside this the node detection was tried to be optimized using
the TensorflowJS API and non max suppression, but it turned out
to be way slower and was appended as a "failed experiment"...