Update constraint detector.
The model is now trained to detect models which did not preserve their
aspect ratio.
Thereby omitting all the image transformation operations,
which were untenable in respect to time needed and memory consumption.
The new approach is nearly agnostic to the number of nodes and
constraints detected and almost accurate enough.

Further improvements should be made to detect horizontal and vertical
constraints which are expected to be a major issue, because the
aspect ratio is very distorted (plus if x1 == x2 or y1 == y2 the
detection is actually skipped...).