Distribute mec2 integration in components
This is kind of a work in progress, because the transition is not
finished completely.
This change gives huge power for clean distinctions of expected behavior
of mec2 properties, e.g. the "base" property of Node may be rendered
as checkbox without having an endless switch statement in MecDetails.

Furthermore an index.js is used to allow cleaner imports.