[wip] HEAVY cleanup - read comment
Refactor project into typescript.
For this rhe react-create-app script is used and the standing app
is patched onto that.
The provided setup is way cleaner than the handmade webpack approach.

For now the App and LeftDrawer is implemented.
Most functionality is not linked yet, but they have to be refactored
into services anyway to be testable.